Update some Nca classes for 14.0.0

+++ tests/LibHac.Tests/Fs/TypeLayoutTests.cs
@@ -476,7 +476,7 @@ public class TypeLayoutTests
     [StructLayout(LayoutKind.Sequential)]
     private struct Int64AlignmentTest
     {
-        public int A;
+        public byte A;
         public Int64 B;
     }
